+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.34.1] - 2022-02-22
+
 ### Changed
 
 - Include assigned variables in `UnusedLocalVariablesRule`.
@@ -838,7 +840,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - False-positives in 13 rules.
 - Logic errors in existing "semantic analysis" that caused scan failures.
 
-[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.34.0...HEAD
+[Unreleased]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.34.1...HEAD
+[0.34.1]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.34.0...v0.34.1
 [0.34.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.33.0...v0.34.0
 [0.33.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.32.0...v0.33.0
 [0.32.0]: https://github.com/integrated-application-development/sonar-delphi/compare/v0.31.1...v0.32.0